+++ src/AnnotatedRoutes/tests/App/Controller/HomeController.php
@@ -8,25 +8,15 @@ use Spiral\Router\Annotation\Route;
 
 class HomeController
 {
-    /**
-     * @Route(route="/", name="index", methods="GET")
-     */
+    #[Route(route: '/', name: 'index', methods: 'GET')]
     public function index(): string
     {
         return 'index';
     }
 
-    /**
-     * @Route(route="/", name="method", methods="POST")
-     */
+    #[Route(route: '/', name: 'method', methods: 'POST')]
     public function method(): string
     {
         return 'method';
-    }
-
-    #[Route(route: '/attribute', name: 'attribute', methods: 'GET', group: 'test')]
-    public function attribute(): string
-    {
-        return 'attribute';
     }
 }