+++ src/Security/src/Rule/ForbidRule.php
@@ -4,14 +4,15 @@ declare(strict_types=1);
 
 namespace Spiral\Security\Rule;
 
-use Spiral\Core\Container\SingletonInterface;
+use Spiral\Core\Attribute\Singleton;
 use Spiral\Security\ActorInterface;
 use Spiral\Security\RuleInterface;
 
 /**
  * Always negative rule.
  */
-final class ForbidRule implements RuleInterface, SingletonInterface
+#[Singleton]
+final class ForbidRule implements RuleInterface
 {
     public function allows(ActorInterface $actor, string $permission, array $context): bool
     {